AlertaDengue/requirements.txt
django
django-bootstrap3
django-leaflet
geojson
gunicorn
numpy
pandas
python-decouple
pytz
psycopg2
sqlalchemy
python-memcached
django-makemessages-xgettext
mock
dbfread
dj-database-url
django-chunked-upload
celery
dj-static
gdal
fiona
rasterio
geopy
